commite

+++ models.py
@@ -1,6 +1,7 @@
-from sqlalchemy import create_engine, String, Integer, DateTime, BIGINT
-from sqlalchemy.orm import DeclarativeBase, Session, Mapped, mapped_column
 from datetime import datetime, timedelta
+
+from sqlalchemy import create_engine, String, Integer, DateTime, BigInteger
+from sqlalchemy.orm import DeclarativeBase, Session, Mapped, mapped_column
 
 # engine = create_engine('sqlite:///taksi_bot.db', echo=True)
 engine = create_engine('postgresql://postgres:1@localhost:5432/taksidb', echo=True)
@@ -16,7 +17,7 @@ class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    user_id: Mapped[int] = mapped_column(BIGINT)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     username: Mapped[str] = mapped_column(String(255), nullable=True)
     chat_id: Mapped[str] = mapped_column(String(255), nullable=True)
     last_permission_granted: Mapped[datetime] = mapped_column(DateTime,
